requirements.txt
python-telegram-bot>=20.0
feedparser>=6.0
beautifulsoup4>=4.12
requests>=2.31
google-generativeai>=0.3
googletrans==4.0.0-rc.1
apscheduler>=3.10
python-dotenv>=1.0
flake8>=6.0
pytest>=7.0
pytest-asyncio>=0.21
